+++ pkg/cmd/kafka/create/create_data.go
@@ -1,4 +1,4 @@
-package kafkautil
+package create
 
 import (
 	"context"
@@ -8,39 +8,6 @@ import (
 	kafkamgmtclient "github.com/redhat-developer/app-services-sdk-go/kafkamgmt/apiv1/client"
 	"github.com/spf13/cobra"
 )
-
-// RegisterNameFlagCompletionFunc adds dynamic completion for the --name flag
-func RegisterNameFlagCompletionFunc(cmd *cobra.Command, f *factory.Factory) error {
-	return cmd.RegisterFlagCompletionFunc("name", func(cmd *cobra.Command, _ []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		var validNames []string
-		directive := cobra.ShellCompDirectiveNoSpace
-
-		conn, err := f.Connection(connection.DefaultConfigSkipMasAuth)
-		if err != nil {
-			return validNames, directive
-		}
-
-		req := conn.API().KafkaMgmt().GetKafkas(f.Context)
-		if toComplete != "" {
-			searchQ := "name like " + toComplete + "%"
-			req = req.Search(searchQ)
-		}
-		kafkas, httpRes, err := req.Execute()
-		if err != nil {
-			return validNames, directive
-		}
-		if httpRes != nil {
-			defer httpRes.Body.Close()
-		}
-
-		items := kafkas.GetItems()
-		for index := range items {
-			validNames = append(validNames, items[index].GetName())
-		}
-
-		return validNames, directive
-	})
-}
 
 // GetCloudProviderCompletionValues returns the list of supported cloud providers for creating a Kafka instance
 // This is used in the cmd.RegisterFlagCompletionFunc for dynamic completion of --provider
@@ -183,89 +150,4 @@ func IsRegionAllowed(region *kafkamgmtclient.CloudRegion, userAllowedInstanceTyp
 		}
 	}
 	return false
-}
-
-// FilterValidTopicNameArgs filters topics from the API and returns the names
-// This is used in for dynamic completion of topic names
-func FilterValidTopicNameArgs(f *factory.Factory, toComplete string) (validNames []string, directive cobra.ShellCompDirective) {
-	validNames = []string{}
-	directive = cobra.ShellCompDirectiveNoSpace
-
-	cfg, err := f.Config.Load()
-	if err != nil {
-		return validNames, directive
-	}
-
-	instanceID, ok := cfg.GetKafkaIdOk()
-	if !ok {
-		return validNames, directive
-	}
-
-	conn, err := f.Connection(connection.DefaultConfigRequireMasAuth)
-	if err != nil {
-		return validNames, directive
-	}
-
-	api, _, err := conn.API().KafkaAdmin(instanceID)
-	if err != nil {
-		return validNames, directive
-	}
-	req := api.TopicsApi.GetTopics(f.Context)
-	if toComplete != "" {
-		req = req.Filter(toComplete)
-	}
-
-	topicRes, _, err := req.Execute()
-	if err != nil {
-		return validNames, directive
-	}
-
-	items := topicRes.GetItems()
-	for _, topic := range items {
-		validNames = append(validNames, topic.GetName())
-	}
-
-	return validNames, directive
-}
-
-// FilterValidConsumerGroupIDs returns the list of consumer group IDs from the API
-func FilterValidConsumerGroupIDs(f *factory.Factory, toComplete string) (validIDs []string, directive cobra.ShellCompDirective) {
-	validIDs = []string{}
-	directive = cobra.ShellCompDirectiveNoSpace
-
-	cfg, err := f.Config.Load()
-	if err != nil {
-		return validIDs, directive
-	}
-
-	instanceID, ok := cfg.GetKafkaIdOk()
-	if !ok {
-		return validIDs, directive
-	}
-
-	conn, err := f.Connection(connection.DefaultConfigRequireMasAuth)
-	if err != nil {
-		return validIDs, directive
-	}
-
-	api, _, err := conn.API().KafkaAdmin(instanceID)
-	if err != nil {
-		return validIDs, directive
-	}
-	req := api.GroupsApi.GetConsumerGroups(f.Context)
-	if toComplete != "" {
-		req = req.GroupIdFilter(toComplete)
-	}
-
-	cgRes, _, err := req.Execute()
-	if err != nil {
-		return validIDs, directive
-	}
-
-	items := cgRes.GetItems()
-	for _, cg := range items {
-		validIDs = append(validIDs, cg.GetGroupId())
-	}
-
-	return validIDs, directive
 }
