+++ pkg/cmd/kafka/create/validator.go
@@ -1,0 +1,136 @@
+package create
+
+import (
+	"errors"
+	"strings"
+
+	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/accountmgmtutil"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/connection"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/remote"
+	kafkamgmtclient "github.com/redhat-developer/app-services-sdk-go/kafkamgmt/apiv1/client"
+	"k8s.io/utils/strings/slices"
+)
+
+type ValidatorInput struct {
+	provider string
+	region   string
+	size     string
+
+	userInstanceTypes []accountmgmtutil.QuotaSpec
+
+	f         *factory.Factory
+	constants *remote.DynamicServiceConstants
+	conn      connection.Connection
+}
+
+func ValidateProviderAndRegion(input *ValidatorInput) error {
+	f := input.f
+	f.Logger.Debug("Validating provider and region")
+	cloudProviders, _, err := input.conn.API().
+		KafkaMgmt().
+		GetCloudProviders(f.Context).
+		Execute()
+
+	if err != nil {
+		return err
+	}
+
+	var selectedProvider kafkamgmtclient.CloudProvider
+
+	providerNames := make([]string, 0)
+	for _, item := range cloudProviders.Items {
+		if !item.GetEnabled() {
+			continue
+		}
+		if item.GetId() == input.provider {
+			selectedProvider = item
+		}
+		providerNames = append(providerNames, item.GetId())
+	}
+	f.Logger.Debug("Validating cloud provider", input.provider, ". Enabled providers: ", providerNames)
+
+	if !selectedProvider.Enabled {
+		providers := strings.Join(providerNames, ",")
+		providerEntry := localize.NewEntry("Provider", input.provider)
+		validProvidersEntry := localize.NewEntry("Providers", providers)
+		return f.Localizer.MustLocalizeError("kafka.create.provider.error.invalidProvider", providerEntry, validProvidersEntry)
+	}
+
+	return validateProviderRegion(input, selectedProvider)
+}
+
+func validateProviderRegion(input *ValidatorInput, selectedProvider kafkamgmtclient.CloudProvider) error {
+	f := input.f
+	cloudRegion, _, err := input.conn.API().
+		KafkaMgmt().
+		GetCloudProviderRegions(f.Context, selectedProvider.GetId()).
+		Execute()
+
+	if err != nil {
+		return err
+	}
+
+	var selectedRegion kafkamgmtclient.CloudRegion
+	regionNames := make([]string, 0)
+	for _, item := range cloudRegion.Items {
+		if !item.GetEnabled() {
+			continue
+		}
+		regionNames = append(regionNames, item.GetId())
+		if item.GetId() == input.region {
+			selectedRegion = item
+		}
+	}
+
+	if len(regionNames) != 0 {
+		f.Logger.Debug("Validating region", input.region, ". Enabled providers: ", regionNames)
+		regionsString := strings.Join(regionNames, ", ")
+		if !selectedRegion.Enabled {
+			regionEntry := localize.NewEntry("Region", input.region)
+			validRegionsEntry := localize.NewEntry("Regions", regionsString)
+			providerEntry := localize.NewEntry("Provider", input.provider)
+			return f.Localizer.MustLocalizeError("kafka.create.region.error.invalidRegion", regionEntry, providerEntry, validRegionsEntry)
+		}
+
+		if err != nil {
+			return err
+		}
+
+		regionInstanceTypes := selectedRegion.GetSupportedInstanceTypes()
+
+		stringTypes := accountmgmtutil.GetInstanceTypes(input.userInstanceTypes)
+		for _, item := range regionInstanceTypes {
+			if slices.Contains(stringTypes, item) {
+				return nil
+			}
+		}
+
+		regionEntry := localize.NewEntry("Region", input.region)
+		userTypesEntry := localize.NewEntry("MyTypes", strings.Join(stringTypes, ", "))
+		cloudTypesEntry := localize.NewEntry("CloudTypes", strings.Join(regionInstanceTypes, ", "))
+
+		return f.Localizer.MustLocalizeError("kafka.create.region.error.regionNotSupported", regionEntry, userTypesEntry, cloudTypesEntry)
+
+	}
+	f.Logger.Debug("No regions found for provider. Skipping provider validation", input.provider)
+
+	return nil
+}
+
+func ValidateSize(input *ValidatorInput) error {
+	amsType, err := accountmgmtutil.PickInstanceType(input.userInstanceTypes)
+	if err != nil {
+		return err
+	}
+	sizes, err := GetValidSizes(input.conn, input.f.Context, input.provider, input.region, &amsType)
+	if err != nil {
+		return err
+	}
+	if !slices.Contains(sizes, input.size) {
+		// TODO error message
+		return errors.New("Whatever") //f.localizer.MustLocalizeError("")
+	}
+	return nil
+}
