+++ pkg/shared/accountmgmtutil/quota.go
@@ -4,6 +4,8 @@ type QuotaType = string
 
 const (
 	// Deprecated by QuotaDeveloperType
-	QuotaTrialType    QuotaType = "eval"
+	QuotaTrialType     QuotaType = "eval"
+	QuotaDeveloperType QuotaType = "developer"
+
 	QuotaStandardType QuotaType = "standard"
 )
